Creates a new base class UmbracoRenderController
This becomes the base class for RenderMvcController and the utlity methods are moved there. This also makes IRenderController discoverable instead of only discovering IRenderMvcController. This all should have been done before v8 release along with removing ContentModel but we are stuck with that. In the meantime, this makes it nicer to hijack routes.
